+++ packages/react/src/screens/index.tsx
@@ -3,7 +3,6 @@ import { usePreludeData, useWidgetRouter } from '@opencx/widget-react-headless';
 import { AnimatePresence } from 'framer-motion';
 import React from 'react';
 import { MotionDiv } from '../components/lib/MotionDiv';
-import { WidgetPortal } from '../components/lib/widget-portal';
 import { ChatScreen } from './chat';
 import { SessionsScreen } from './sessions';
 import { WelcomeScreen } from './welcome';
@@ -17,55 +16,52 @@ export function RootScreen() {
   } = useWidgetRouter();
 
   return (
-    <>
-      <WidgetPortal.Container />
-      <div className="relative bg-background size-full">
-        <AnimatePresence mode="wait">
-          {(() => {
-            switch (screen) {
-              case 'welcome':
-                return (
-                  <MotionDiv
-                    key={screen}
-                    fadeIn="right"
-                    className="size-full"
-                    snapExit
-                  >
-                    <WelcomeScreen />
-                  </MotionDiv>
-                );
+    <div className="relative bg-background size-full">
+      <AnimatePresence mode="wait">
+        {(() => {
+          switch (screen) {
+            case 'welcome':
+              return (
+                <MotionDiv
+                  key={screen}
+                  fadeIn="right"
+                  className="size-full"
+                  snapExit
+                >
+                  <WelcomeScreen />
+                </MotionDiv>
+              );
 
-              case 'sessions':
-                return (
-                  <MotionDiv
-                    key={screen}
-                    fadeIn="right"
-                    className="size-full"
-                    snapExit
-                  >
-                    <SessionsScreen />
-                  </MotionDiv>
-                );
+            case 'sessions':
+              return (
+                <MotionDiv
+                  key={screen}
+                  fadeIn="right"
+                  className="size-full"
+                  snapExit
+                >
+                  <SessionsScreen />
+                </MotionDiv>
+              );
 
-              case 'chat':
-                return (
-                  <MotionDiv
-                    key={screen}
-                    fadeIn="right"
-                    className="size-full"
-                    snapExit
-                  >
-                    <ChatScreen />
-                  </MotionDiv>
-                );
-              default: {
-                isExhaustive(screen, RootScreen.name);
-                return null;
-              }
+            case 'chat':
+              return (
+                <MotionDiv
+                  key={screen}
+                  fadeIn="right"
+                  className="size-full"
+                  snapExit
+                >
+                  <ChatScreen />
+                </MotionDiv>
+              );
+            default: {
+              isExhaustive(screen, RootScreen.name);
+              return null;
             }
-          })()}
-        </AnimatePresence>
-      </div>
-    </>
+          }
+        })()}
+      </AnimatePresence>
+    </div>
   );
 }